@@ -3,15 +3,15 @@ description: Shows examples for the use of the flutter_animator_widgets package.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^0.1.2
-  flutter_animator_widgets:
-    path: ../
+  flutter_animator_widgets: ^1.0.2
+#    path: ../
 
 dev_dependencies:
   flutter_test:
