@@ -1,17 +1,16 @@
 name: flutter_animator_widgets
 description: Awesome customizable animated widgets. Widgets are build with the flutter_animator package. (you should check it out ;))
-version: 1.0.1
-author: Sharp3dges <sharp3dges@outlook.com>
+version: 1.0.2
 homepage: https://github.com/sharp3dges/animator_widgets
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_animator: ^2.0.3
-  vector_math: ^2.0.8
+  flutter_animator: ^3.1.1
+  vector_math: ^2.1.0
 
 dev_dependencies:
   flutter_test:
